♻️ refactor: 도메인별 controller, service 디렉토리 추가
적용된 도메인들: booking, place, reservation, util

Issue Resolved: #59
