+++ back/src/domains/booking/booking.module.ts
@@ -4,7 +4,7 @@ import { Module } from '@nestjs/common';
 import { AuthModule } from '../../auth/auth.module';
 import { EventModule } from '../event/event.module'; // Import EventModule
 
-import { BookingService } from './booking.service';
+import { BookingService } from './service/booking.service';
 
 @Module({
   imports: [EventModule, AuthModule],
