+++ back/src/domains/reservation/controller/reservation.controller.spec.ts
@@ -1,7 +1,8 @@
 import { Test, TestingModule } from '@nestjs/testing';
 
+import { ReservationService } from '../service/reservation.service';
+
 import { ReservationController } from './reservation.controller';
-import { ReservationService } from './reservation.service';
 
 describe('ReservationController', () => {
   let controller: ReservationController;
